+++ includes/ClicShopping/Apps/Payment/Stripe/Module/Payment/ST.php
@@ -226,7 +226,7 @@ pre_confirmation_check()
 
       if (!isset($stripe_payment_intent_id)) {
         $description = STORE_NAME . ' - Order date time : ' . date('Y-m-d H:i:s');
-        $token = isset(  $_POST['stripeToken'] ) ?  $_POST['stripeToken'] : [];
+        $token = $_POST['stripeToken'] ?? [];
 
         $params = [
             'amount' => $total_amount,