+++ includes/ClicShopping/Apps/Payment/Stripe/Sites/ClicShoppingAdmin/Pages/Home/Actions/Configure/Process.php
@@ -13,7 +13,7 @@
   use ClicShopping\OM\Registry;
 
   class Process extends \ClicShopping\OM\PagesActionsAbstract  {
-    public function execute() 
+    public function execute()
     {
       $CLICSHOPPING_MessageStack = Registry::get('MessageStack');
       $CLICSHOPPING_Stripe = Registry::get('Stripe');