+++ includes/ClicShopping/Sites/ClicShoppingAdmin/CkEditor4.php
@@ -71,7 +71,7 @@
 
         return $connector;
       } else {
-        HTTP::redirect();
+        CLICSHOPPING::redirect();
       }
     }