+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 version 3.434
-Does not allow to access elfinder without session connexion
 Brace function updated
+Fix db syntaxe for marketplace
 
 --------------------
 version 3.433 and 3.4331